Pyannote_Diarization_Separation_Onnx/requirements.txt
# Core dependencies
pyannote.audio>=3.0.0
torch>=2.0.0
torchaudio>=2.0.0

# Audio processing
soundfile>=0.12.0
librosa>=0.10.0

# ONNX support
onnxruntime>=1.16.0
# Use onnxruntime-gpu instead if you have CUDA GPU
# onnxruntime-gpu>=1.16.0

# Utilities
numpy>=1.24.0
scipy>=1.10.0
einops>=0.7.0

# Optional: For visualization and analysis
matplotlib>=3.7.0
pandas>=2.0.0
